refactor : user에서 info 분리
