+++ src/modules/user/user.controller.ts
@@ -3,9 +3,9 @@ import { Response } from 'express';
 import { UserService } from './user.service';
 import { UserDto } from './dto/user.dto';
 import { ApiTags, ApiOperation } from '@nestjs/swagger';
-import { InfoDto } from './dto/info.dto';
+import { InfoDto } from '../myinfo/dto/info.dto';
 
-@ApiTags('signup')
+@ApiTags('/signup')
 @Controller()
 export class UserController {
   constructor(private readonly userService: UserService) {}
@@ -36,41 +36,5 @@ export class UserController {
     }
 
     return res.status(201).json({ message: '회원가입 완료' });
-  }
-
-  /**
-   * 내 정보/북마크/스터디/게시물 조회
-   * @param userDto 
-   * @returns 
-   */
-  @ApiOperation({
-    summary: '내 정보/북마크/스터디/게시물 조회 API',
-    description: '내 정보, 관심게시물, 참여스터디, 작성게시물 조회 API입니다.',
-  })
-  @Get('/user/my-info')
-  async getUserInfo(@Body() infoDto: InfoDto) {
-    try {
-      
-      if (!infoDto.userId) throw new HttpException('로그인이 필요한 서비스 입니다',HttpStatus.BAD_REQUEST)
-      console.log(infoDto);
-
-      const methodMap = {
-        bookmarks : 'getBookmarks',
-        studylists : 'getStudylists',
-        posts : 'getPosts'
-      }
-
-      const methodName = methodMap[infoDto.name] || 'getUserInfo'
-
-      const result = await this.userService[methodName](infoDto)
-     
-      console.log('결과값 > ', result)
-      return result;
-      
-    } catch (error) {
-      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
-    }
-  }
-
-  
+  }  
 }
