+++ src/modules/user/user.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/database/prisma/prisma.service';
 import { UserDto } from './dto/user.dto';
-import { InfoDto } from './dto/info.dto';
 
 @Injectable()
 export class UserService {
@@ -64,95 +63,4 @@ export class UserService {
       console.log(error);
     }
   }
-
-  /**
-   * 내 정보 조회
-   * @param userDto 
-   * @returns users skills 조회
-   */
-  async getUserInfo(infoDto: InfoDto) {
-    console.log('getUserInfo >>> ', infoDto.userId);
-    const user = await this.prisma.$queryRaw`
-    select a.userId,a.userName,a.userNickName,
-    a.position,GROUP_CONCAT(DISTINCT b.skill) as skills from
-    users a join
-    skills b on
-    a.userId = b.userId
-    where a.userId = ${Number(infoDto.userId)}
-    `;
-    return user;
-  }
-
-  /**
-   * 내 북마크 조회
-   * @param userDto 
-   * @returns 게시글번호,게시물제목,게시글작성자
-   */
-  async getBookmarks(infoDto: InfoDto) {
-    console.log(`북마크`);
-
-    const user = await this.prisma.$queryRaw
-    `select b.postId,b.postTitle,b.post_userId,
-    a.userId, c.userNickName
-    from 
-    bookmarks a
-    join
-    posts b 
-    on a.postId = b.postId
-    join 
-    users c 
-    on b.post_userId = c.userId
-    where a.userId = ${Number(infoDto.userId)}`;
-
-    return user
-  }
-
-  /**
-   * 내 참여 스터디 조회
-   * @param userDto 
-   * @returns 게시글번호,게시글제목,게시글타입,게시글작성자
-   */
-  async getStudylists(infoDto: InfoDto) {
-    console.log(`스터디`);
-    const user = await this.prisma.$queryRaw
-    `select b.postId,b.postTitle,b.postType,
-    b.post_userId,a.userId,c.userNickName
-    from
-    studylists a
-    join
-    posts b 
-    on a.postId = b.postId
-    join 
-    users c 
-    on b.post_userId = c.userId
-    where a.userId = ${Number(infoDto.userId)}`;
-
-    return user
-
-  }
-
-  /**
-   * 내 작성 게시물 조회
-   * @param userDto 
-   * @returns 게시글번호,게시글제목,게시글타입 orderby 글 작성 최신순
-   */
-  async getPosts(infoDto: InfoDto) {
-    console.log(`게시물`);
-    const user = await this.prisma.posts.findMany({
-      where:{
-        post_userId:Number(infoDto.userId)
-      },
-      select:{
-        postId:true,
-        postTitle:true,
-        postType:true,
-      },
-      orderBy:{
-        createdAt:'desc'
-      }
-    })
-
-    return user
-  }
-
 }
